[fix]create keyword - transfer: refactoring
